# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Reactor
  version: 1.26.10-beta8
  license:
    name: BUSL-1.1
    url: https://mariadb.com/bsl11/
  description: |
    Auth, storage, functions, sites, health, and platform routes.
    Data stays on PostgREST at /data/v1.
servers:
  - url: http://127.0.0.1:18000
    description: Local Compose
...
paths:
...
  /platform/v1/projects/{ref}:
    patch:
      tags: [platform]
      summary: Point a project at a dedicated database
      security: [{operator: []}]
      parameters:
        - $ref: "#/components/parameters/Ref"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [database_url]
              properties:
                database_url: {type: string}
      responses:
        "204":
          description: Updated
...
  /auth/v1/verify-email:
    post:
      tags: [auth]
      summary: Confirm an email with a link token or a code
      parameters:
        - $ref: "#/components/parameters/ApiKey"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/verify-email/send:
    post:
      tags: [auth]
      summary: Send a new confirmation message
      parameters:
        - $ref: "#/components/parameters/ApiKey"
      responses:
        "200":
          description: Always ok
...
  /console/v1/projects/{pref}/auth:
    get:
      tags: [console]
      summary: Project auth settings
      responses:
        "200":
          description: Verification and 2FA flags
        "403":
          $ref: "#/components/responses/Error"
    put:
      tags: [console]
      summary: Update project auth settings
      responses:
        "200":
          description: Updated flags
        "403":
          $ref: "#/components/responses/Error"
...
components:
...
  parameters:
    Ref:
      name: ref
      in: path
      required: true
      schema: {type: string}
...
    ApiKey:
      name: Authorization
      in: header
      required: false
      description: Bearer anon key, or the project host
      schema: {type: string}
...
  schemas:
...
    User:
      type: object
      properties:
        id: {type: string}
        email: {type: string}
    Session:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        user:
          $ref: "#/components/schemas/User"
...
    ErrorBody:
      type: object
      properties:
        error: {type: string}
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  securitySchemes:
    operator:
      type: http
      scheme: bearer
      description: Operator token
    service:
      type: http
      scheme: bearer
      description: Project service key
    user:
      type: http
      scheme: bearer
      description: User access token
    project:
      type: http
      scheme: bearer
      description: Anon key, user access token, or service key